Integración clase cliente y clienteBean
Borramos el archivo cliente bean y lo pusimos todo en el único archivo cliente.
